backend/requirements.txt
Django>=5.0
djangorestframework
djangorestframework-simplejwt
django-filter
psycopg2-binary
celery
redis
drf-spectacular
pytest==7.1.3
pytest-django
factory_boy
pytest-lazy-fixture
django-celery-beat